KVM: ARM: vgic: Keep track of the physical ID using the vgic
Add a cpuid field to the vgic_cpu structure so that we know which vgic
instance (tied to a specific vcpu instance) is currently running on a
physical CPU, and which physical CPU it's running on.

This gives us the possibility of signaling such a physical CPU directly
if it has any free list registers without having to do a hardware IPI
and have the guest CPU take a round-trip to the host to program a single
list register.

Signed-off-by: Christoffer Dall <cdall@cs.columbia.edu>